bug9302: shut the warning: argument 1 ... range exceeds maximum object size warning
warning: argument 1 range [18446744071562067968, 18446744073709551615] exceeds maximum object size 9223372036854775807

I don't like this fix, so if you're reading the code and have a better
idea - you're welcome. Tried to make everything unsigned to avoid
(int) conversion, but it effectively means all fields in all TRC
structures should change type and this does not look safe.

gcc (Ubuntu 7.3.0-27ubuntu1~18.04) 7.3.0

+++ tools/trc/lib/update.c
@@ -4941,7 +4941,7 @@ trc_update_gen_test_wilds_fss(unsigned int db_uid,
                               trc_update_args_groups *wildcards,
                               uint64_t flags)
 {
-    int          ids_count;
+    size_t       ids_count;
     int          res_id;
     int          i = 0;
     int          j = 0;
@@ -4976,7 +4976,7 @@ trc_update_gen_test_wilds_fss(unsigned int db_uid,
 
     gettimeofday(&tv_before_gen_fss, NULL);
 
-    for (res_id = 1; res_id <= ids_count; res_id++)
+    for (res_id = 1; res_id <= (int)ids_count; res_id++)
     {
         memset(&args_groups, 0, sizeof(args_groups));
         SLIST_INIT(&args_groups);
@@ -5010,7 +5010,7 @@ trc_update_gen_test_wilds_fss(unsigned int db_uid,
 
     wild_prbs = calloc(ids_count, sizeof(*wild_prbs));
 
-    for (res_id = 1; res_id <= ids_count; res_id++)
+    for (res_id = 1; res_id <= (int)ids_count; res_id++)
     {
         p_id = res_id - 1;
         wild_prbs[p_id].set_max = 10;
@@ -5083,7 +5083,7 @@ trc_update_gen_test_wilds_fss(unsigned int db_uid,
 
     if (flags & TRC_UPDATE_EXT_WILDS)
     {
-        for (i = 0; i < ids_count; i++)
+        for (i = 0; i < (int)ids_count; i++)
         {
             for (j = 0; j < wild_prbs[i].sol_num; j++)
             {
@@ -5125,7 +5125,7 @@ trc_update_gen_test_wilds_fss(unsigned int db_uid,
     }
 
     /* Insert generated wildcards in TRC DB */
-    for (i = 0; i < ids_count; i++)
+    for (i = 0; i < (int)ids_count; i++)
     {
         for (j = 0; j < wild_prbs[i].sol_num; j++)
         {
